Disable logging by default
Starting on `bunyan@1.2.2`, it is possible to (again) disable logging by
defining a level higher than `bunyan.FATAL`. This is now the default as
it more accurately represents what the inexistence of the `DEBUG` env
var has on logging when using the `debug` module - no output is
generated by loggers.